Issue #20: Added workout update code topic question removing or creation:
When added or deleted topic question, then the following properties in `UserTopicWorkout` (see the prisma schema) table should be updated (if the corresponding record, byt topicId, is existed):

- questionsOrder
- questionResults
- stepIndex

If the question has been added, then add the question id to the end of `questionsOrder` property (with a space delimiter).

If the question has been removed, then find the question id index in the (space delimited) `questionsOrder` field, and if it found, then remove the id, remove the `questionResults` (json packed array) at this index, and decrease `stepIndex` if found index is lower than `stepIndex` value.

The logic is in the following modules:

- `src/features/questions/actions/addNewQuestion.ts`
- `src/features/questions/actions/deleteQuestion.ts`

Tests added in the `src/features/questions/actions/__tests__` folder.

Side effects: added `@jest/globals`, updated prisma dependency.